Use stdint.h
When possible and relevant, switch to known-sized integer types. We did
not do this for `char` because `char` is either signed or unsigned, and
none of stdint.h can handle that. We did not do this in cases where the
libc function mentions `int` because that is a defined interface that we
do not want to violate.